Struct handlebars::Decorator

pub struct Decorator<'a> { /* fields omitted */ }

Render-time Decorator data when using in a decorator definition

Methods

impl<'a, 'b> Directive<'a>

pub fn name(&self) -> &str

Returns helper name

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn params(&self) -> &Vec<ContextJson>

Returns all helper params, resolved within the context

pub fn param(&self, idx: usize) -> Option<&ContextJson>

Returns nth helper param, resolved within the context

pub fn hash(&self) -> &BTreeMap<String, ContextJson>

Returns hash, resolved within the context

pub fn hash_get(&self, key: &str) -> Option<&ContextJson>

Return hash value of a given key, resolved within the context

pub fn template(&self) -> Option<&Template>

Returns the default inner template if any