Struct handlebars::RenderContext

pub struct RenderContext<'reg, 'rc> { /* fields omitted */ }

The context of a render call

This context stores information of a render and a writer where generated content is written to.

Implementations

impl<'reg: 'rc, 'rc> RenderContext<'reg, 'rc>

pub fn new(root_template: Option<&'reg String>) -> RenderContext<'reg, 'rc>

Create a render context from a Write

pub fn push_block(&mut self, block: BlockContext<'reg>)

Push a block context into render context stack. This is typically called when you entering a block scope.

pub fn pop_block(&mut self)

Pop and drop current block context. This is typically called when leaving a block scope.

pub fn block(&self) -> Option<&BlockContext<'reg>>

Borrow a reference to current block context

pub fn block_mut(&mut self) -> Option<&mut BlockContext<'reg>>

Borrow a mutable reference to current block context in order to modify some data.

pub fn context(&self) -> Option<Rc<Context>>

Get the modified context data if any

pub fn set_context(&mut self, ctx: Context)

Set new context data into the render process. This is typically called in decorators where user can modify the data they were rendering.

pub fn evaluate(
&self,
context: &'rc Context,
relative_path: &str
) -> Result<ScopedJson<'reg, 'rc>, RenderError>

Evaluate a Json path in current scope.

Typically you don’t need to evaluate it by yourself. The Helper and Decorator API will provide your evaluated value of their parameters and hash data.

pub fn get_partial(&self, name: &str) -> Option<&&Template>

Get registered partial in this render context

pub fn set_partial(&mut self, name: String, partial: &'reg Template)

Register a partial for this context

pub fn remove_partial(&mut self, name: &str)

Remove a registered partial

pub fn is_current_template(&self, p: &str) -> bool

Test if given template name is current template.

pub fn register_local_helper(
&mut self,
name: &str,
def: Box<dyn HelperDef + 'rc>
) -> Option<Rc<dyn HelperDef + 'rc>>

Register a helper in this render context. This is a feature provided by Decorator where you can create temporary helpers.

pub fn unregister_local_helper(&mut self, name: &str)

Remove a helper from render context

pub fn get_local_helper(&self, name: &str) -> Option<Rc<dyn HelperDef + 'rc>>

Attempt to get a helper from current render context.

pub fn get_current_template_name(&self) -> Option<&'reg String>

Returns the current template name. Note that the name can be vary from root template when you are rendering from partials.

pub fn set_current_template_name(&mut self, name: Option<&'reg String>)

Set the current template name.

pub fn get_root_template_name(&self) -> Option<&'reg String>

Get root template name if any. This is the template name that you call render from Handlebars.

pub fn is_disable_escape(&self) -> bool

Get the escape toggle

pub fn set_disable_escape(&mut self, disable: bool)

Set the escape toggle. When toggle is on, escape_fn will be called when rendering.